requirements.txt
# App dependencies
azure-functions
azure-storage-blob
opencv-python
numpy
Pillow
huggingface-hub
torch
transformers
timm
# TODO: Remove ultralytics once DETR rollout is stable.
ultralytics

# Dev/test dependencies
pytest
pytest-cov
ruff
mypy
